REMOVE ONLINE SERVICE COMPLETELY
