+++ content_store/lib/adapters/stubs/server_ui_stubs.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class ServerControl extends StatelessWidget {
+  const ServerControl({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox.shrink();
+  }
+}
+
+class ServerSettings extends StatelessWidget {
+  const ServerSettings({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox.shrink();
+  }
+}
