+++ content_store/lib/content_store.dart
@@ -1,9 +1,12 @@
 /// Content Store.
 library content_store;
 
+export 'adapters/stubs/server_ui_stubs.dart';
 export 'db_service/builders/get_db_reader.dart';
 export 'db_service/builders/get_media_text.dart';
+export 'db_service/builders/get_media_uri.dart';
 export 'db_service/builders/get_notes.dart';
+export 'db_service/builders/get_preview_uri.dart';
 export 'db_service/builders/get_store_updater.dart';
 export 'db_service/builders/w3_get_collection.dart';
 export 'db_service/builders/w3_get_media.dart';
@@ -12,8 +15,4 @@ export 'db_service/models/store_updter_ext_store.dart';
 export 'db_service/widgets/broken_image.dart';
 export 'db_service/widgets/shimmer.dart';
 export 'extensions/ext_cl_medias.dart';
-export 'online_service/builders/get_media_uri.dart';
-export 'online_service/builders/get_preview_uri.dart';
-export 'online_service/widgets/server_control.dart';
-export 'online_service/widgets/server_settings.dart';
 export 'storage_service/widgets/storage_monitor.dart' show StorageMonitor;
