--- content_store/lib/online_service/builders/get_server.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../models/server.dart';
-import '../providers/server.dart';
-
-class GetServer extends ConsumerWidget {
-  const GetServer({
-    required this.builder,
-    super.key,
-    this.errorBuilder,
-    this.loadingBuilder,
-  });
-  final Widget Function(Server server) builder;
-  final Widget Function(Object, StackTrace)? errorBuilder;
-  final Widget Function()? loadingBuilder;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final server = ref.watch(serverProvider);
-    return builder(server);
-  }
-}
