--- content_store/lib/online_service/providers/downloader_status.dart
@@ -1,7 +1,0 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-import '../models/downloader_status.dart';
-
-final downloaderStatusProvider = StateProvider<DownloaderStatus>((ref) {
-  return const DownloaderStatus();
-});
